requirements.txt
arabic-reshaper==3.0.0
asn1crypto==1.5.1
astroid==3.3.8
black==25.1.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cryptography==45.0.4
cssselect2==0.8.0
cycler==0.12.1
dill==0.3.9
Flask==3.1.1
Flask-Mail==0.10.0
fonttools==4.54.1
html5lib==1.1
idna==3.10
iniconfig==2.0.0
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mccabe==0.7.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.1.3
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
pygame==2.6.1
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pylint==3.3.4
pyparsing==3.2.0
pypdf==5.6.1
pytest==8.3.4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
svglib==1.5.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tomlkit==0.13.2
tzdata==2024.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.17
